+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add "fillFieldWithinDefaultTimeoutUsingJavaScript" step that allows filling input fields within a default timeout using JavaScript
 - Add "activateIFrame" that allows switching the currently active frame to a given iFrame
 - Add spinned version of MinkContext's "assertElementNotContainsText" method
+- Add "HelpUtils" trait for helper functions that can be useful for defining own steps
+- Add "assertJavascriptExpression" method to the "HelpUtils" trait wrapping the wait method of Mink's Session 
  
 ### Changed
 - Increase the default waiting interval for the SpinnedMinkSteps from 5 seconds to 15 seconds